+++ src/main/java/com/projetoCrud/projetoCrud01/services/PessoaServices.java
@@ -7,8 +7,8 @@ import com.projetoCrud.projetoCrud01.model.Pessoa;
 public interface PessoaServices {
 	
 	List<Pessoa> findAll();
-	Pessoa findById(long id);
+	Pessoa findById(int id);
 	Pessoa save(Pessoa pessoa);
-	Pessoa deleteById(long id);
+	Pessoa deleteById(int id);
 
 }
